Overhaul projectiles.
Multiplayer testing not done because studio is being weird

* PlayerAbilities.lua:
spawns `Usage`s of `Abilities` which create `Components` (currently called projectiles that are associated with the usage. The `ClientUse` function of the Ability will create the components; `Projectile:Fire` will associate itself with the usage passed to it, but instances which determine if a usage is still active should be owned by it by calling `Usage:OwnInstance`. Once all owned instances are destroyed, the client notifies the server that the usage can be deleted.
Projectiles have sensible default control functions, and a small feature is that `Projectile.new` can recognize presets such as AoE rather than raycasts.
Usages have a client-controlled ID, and components they spawn have client-owned IDs relative to the usage.
See Weapons' use of Abilities at the bottom of the file for an example of creating an ability.

* ClientStates.lua:
This module is barebones right now. Prior to using an ability, the client can set a state that is synced with the server. Only one state may be active at a time, but the states can relate to eachother and e.g. prevent going to one state from a specific other state (this feature is not finished). On the client or server, states can propagate callbacks.

* Util/PlayerRegistry.lua
A structure that constructs an entry in a table for each player in the game. It syncs the list of players to players currently in game. Example usage can be seen in PlayerAbilities.lua

* Fix the `PlayerJoined` and `PlayerLeaved` game events. They didn't really work at all before. See changes in PlayerLib.lua

* Game events can now pass a parameter which specifies the subject that is the noun being verbed. E.g. `PlayerJoined` will pass along the player which joined now.

* StraightShot.lua:
The ability in this file is just a placeholder